Correctly localize debug schema attributes
